[FIX] Migrating ``required`` from model to view
While migrating ``fieldservice_repair`` (https://github.com/OCA/field-service/pull/1209) some unit tests were failing because this module adds ``required`` condition on two new fields on ``fms.equipment`` model.
To avoid this, the best solution was to migrate the required to the view as we cannot force the installation order on odoo.